Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -2,6 +2,6 @@
 int main(){
     float a,b;
     scanf("%f %f",&a ,&b);
-    int Product=a*b;
+    float Product=a*b;
     printf("Product: %.2f",Product);
 }
